Only call DeleteItem if AcceptItem succeeds for each item

+++ broker/lms/lms_adapter_ncip.go
@@ -166,6 +166,9 @@ func (l *LmsAdapterNcip) AcceptItem(
 }
 
 func (l *LmsAdapterNcip) DeleteItem(itemId string) error {
+	if l.config.AcceptItemEnabled != nil && !*l.config.AcceptItemEnabled {
+		return nil
+	}
 	arg := ncip.DeleteItem{
 		ItemId: ncip.ItemId{ItemIdentifierValue: itemId},
 	}
